backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Git and Repository Analysis
GitPython==3.1.41
radon==6.0.1

# Graph and Data Processing
networkx==3.2.1
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# CORS
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
